agentic_work/17_agent_as_tool/pyproject.toml
[project]
name = "17-agent-as-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
